WpfPlot as CustomControl
- update ScottPlot5-wpf.slnf to add missed ScottPlot.OpenGL project
- WpfPlot use now a template and is based on System.Windows.Controls.Control
