+++ src/ScottPlot5/ScottPlot5 Controls/ScottPlot.WPF/Themes/Generic.xaml
@@ -1,0 +1,17 @@
+<ResourceDictionary
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="clr-namespace:ScottPlot.WPF"
+    xmlns:skiaPreview="clr-namespace:SkiaSharp.Views.WPF"
+    >
+
+    <Style TargetType="{x:Type local:WpfPlot}">
+        <Setter Property="Template">
+            <Setter.Value>
+                <ControlTemplate TargetType="{x:Type local:WpfPlot}">
+                    <skiaPreview:SKGLElement x:Name="PART_SKElement" />
+                </ControlTemplate>
+            </Setter.Value>
+        </Setter>
+    </Style>
+</ResourceDictionary>
