--- src/ScottPlot5/ScottPlot5 Controls/ScottPlot.WPF/WpfPlot.xaml
@@ -1,24 +1,0 @@
-﻿<UserControl x:Class="ScottPlot.WPF.WpfPlot"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:skia="clr-namespace:SkiaSharp.Views.WPF;assembly=SkiaSharp.Views.WPF"
-             xmlns:skiaPreview="clr-namespace:SkiaSharp.Views.WPF"
-             xmlns:local="clr-namespace:ScottPlot.WPF"
-             mc:Ignorable="d" 
-             d:DesignHeight="450" d:DesignWidth="800"
-             Focusable="True"
-             KeyDown="OnKeyDown"
-             KeyUp="OnKeyUp"
-             >
-    <skiaPreview:SKGLElement 
-        x:Name="SKElement" 
-        PaintSurface="OnPaintSurface" 
-        MouseDown="OnMouseDown" 
-        MouseUp="OnMouseUp" 
-        MouseMove="OnMouseMove" 
-        MouseWheel="OnMouseWheel"
-    />
-</UserControl>
-
