+++ src/ScottPlot5/ScottPlot5-wpf.slnf
@@ -4,6 +4,7 @@
     "projects": [
       "..\\Shared\\ScottPlotTests\\ScottPlotSharedTests.csproj",
       "..\\Shared\\ScottPlot\\ScottPlotShared.shproj",
+      "ScottPlot5 Controls\\ScottPlot.OpenGL\\ScottPlot.OpenGL.csproj",
       "ScottPlot5 Controls\\ScottPlot.WPF\\ScottPlot.WPF.csproj",
       "ScottPlot5 Cookbook\\ScottPlot Cookbook.csproj",
       "ScottPlot5 Sandbox\\Sandbox.WPF\\Sandbox.WPF.csproj",
